We haz spheres.
Okay, just an explanation of the changes.

C++ doesn't allow for virtual static data, so I had to do a
hack to get it.  But data is now loaded in a single method, using
class data.

Also, I loaded up an actual sphere to be drawn to the screen,
thus the title for the commit.